+++ src/ui/folder_select.py
@@ -93,6 +93,10 @@ class FolderSelectWidget(QWidget):
 
     """navigation section: clicking in the tree, entering the path in the edit, or using back & forward buttons"""
 
+    def force_set_directory(self, str):
+        self.folder_edit.setText(str)
+        self.set_directory_upon_edit()
+
     def set_directory_upon_edit(self):
         # get the path
         path = QFileInfo(self.folder_edit.text()).absoluteFilePath()